Allow keepalives to be enabled for REv2 gRPC channels
Bazel currently doesn't make use of gRPC's keepalive options, meaning
that builds may get stuck unnecessarily when connected to a server that
disappears off the network. It may well be the case that reconnecting to
an alternative server allows the build to proceed.

Considering that Bazel's RPC rate is already relatively high, there is
generally little harm in enabling keepalives from a performance point of
view, especially if we don't send them when idle (i.e., the default).

This change introduces two new configuration options,
--remote_keepalive_time (the interval) and --remote_keepalive_timeout.

Fixes: https://github.com/bazelbuild/bazel/issues/7085