Add transition support for Apple bitcode
Embedding bitcode to the watchOS app binary is mandatory if you want to
submit the app to the App Store. Previously it wasn't possible to enable
bitcode just for the watchOS app but not for the iOS app that bundles
it.

This patch adds a new `--watchos_force_bitcode` build option --- when
enabled, it will set the bitcode mode to "embedded" for the produced
watchOS binaries, regardless of the value of the global
`--apple_bitcode` option.

Closes https://github.com/bazelbuild/bazel/issues/11278.

RELNOTES: Apple bitcode for watchOS binaries can now be forced embedded
with the new flag `--watchos_force_bitcode`.