Prohibit injecting tree artifact children as files. They should only be injected via injectDirectory.
This is consistent with the check in ActionExecutionValue which ensures that tree artifact children are not stored outside of their containing TreeArtifactValue.

This change allows for a followup optimization in the way we handle chmod calls on outputs - tree artifact children were the only case where a file could be injected and still passed to setPathReadOnlyAndExecutable.

PiperOrigin-RevId: 319844401