Fix newly introduced bug where an empty test_suite.tests attribute was incorrectly not causing the test_suite.$implicit_tests attribute to get set.
The test_suite.$implicit_tests attribute is the implementation of the latter paragraph in
the documentation of the test_suite.tests attribute
(https://docs.bazel.build/versions/3.3.0/be/general.html#test_suite); this bug
here was causing that feature to be broken in the "or empty" case.

This bug was introduced in commit a46ad6.

RELNOTES: None
PiperOrigin-RevId: 320109724