Limit warnings about shared actions with tree artifact input to 1 per build.
Bazel does not handle shared actions with tree artifact inputs properly as it
is overly permissive in such case. This means that certain builds, which should
fail, will be allowed, which can result with non-deterministic results. We
recently added a warning about that, which turned out to be overly chatty.

Restrict the warning to be issued once per build only and all of the
information to the log.

PiperOrigin-RevId: 320266872