bazel packages: record load graph (DAG over Modules) in Module and Package
This is a preparatory step for a new, rich query output format that will provide
all the information needed to implement stardoc as a client of blaze, thus allowing
us to delete the fakebuildapi implementations and the starlarkbuildapi abstractions.

In some situations this could increase live heap usage by causing a package
to keep its loaded modules live for longer. Specifically, if packages P and Q
both load module M, then M is changed and package Q is reloaded, the stale
package P remains live, and it holds a reference to the stale M, whereas before
it did not. The next time P is loaded, the stale P and M are evicted.
Note that today, P already keeps alive any modules required by rules instantiated
by P, so an additional cost will be incurred only if P uses functions ("macros")
or data defined in M, but no rules. I expect the effect to be quite marginal.

A number of opportunities for simplification (and memory saving) have been
identified and will be acted on in a follow-up.

Also:
- define BazelModuleContext.of helper function.

RELNOTES: N/A
PiperOrigin-RevId: 320283510