Refactor interaction between ActionExecutionFunction and SkyframeActionExecutor.
A lot of stuff is being passed back and forth. Let SkyframeActionExecutor be responsible for managing the objects it needs (for example updating the action filesystem context and closing the ActionExecutionContext).

The MetadataConsumer for input discovery is changed to one that always throws, since we don't expect any outputs during input discovery.

PiperOrigin-RevId: 320707537