Have WorkspaceFileFunctionTests go through proper skyframe evaluations
We had a mix of mocked-up env and "proper" evaluation. Settle on the latter
since it's less code and should be easier to maintain.

Motivated by attempting to reduce/remove Package#getEvents, which should save
some memory and API surface area. It'll also force us to test that the events
that are generated are actually reported. That'll come in later changes.

FakeFileValue is no longer necessary in WorkspaceFileFunctionTests, so move it
to where it's needed and make it private.

PiperOrigin-RevId: 320759614