Starlark: Cache descriptor resolution in getattr
Even with descriptor cache, hash map lookup is quite expensive.

This is especially hurtful when getattr is invoked for structs:
* we lookup for descriptor before checking if class implements `ClassObject`
* we don't cache negative lookups (and we probably shouldn't because
  structs can have bazillion distinct fields)

Benchmark:

```
def test():
    l = []
    for i in range(10):
        print(i)
        for j in range(1000):
            for k in range(2000):
                a = l.append
                a = l.append
                a = l.append
                a = l.append
                a = l.append
                a = l.append
                a = l.append
                a = l.append

test()
```

Results are:

```
A: N=18, r=10.716+-0.358
B: N=18, r=8.063+-0.386
B/A: 0.752
```

This implementation caches the resolution inside the AST.  With
bytecode interpreter the same caching can be done in the bytecode
object.

The similar optimization can be applied to `Starlark.fastcall`, but
before doing that I'd like to get feedback on this.