Remove OutputStore#injectedFiles.
This set was previously used to track which files were injected with the goal of skipping chmod calls on injected files. After recent changes, there is no longer any code path that attempts to call setPathReadOnlyAndExecutable on an artifact that was already injected in a prior call.

This change also removes duplicate chmod calls for local tree artifact files and adds tests counting chmod calls. Because setPathReadOnlyAndExecutable always stats the path to check if it's a file, this also reduces the number of stats on tree artifact files.

Fun fact: OutputStore is now down to just 2 maps - at one point it was up to 5 data stores (4 maps & 1 set).

PiperOrigin-RevId: 321060966