Report workspace file events as they happen
This is closer to how events for proper packages are reported (immediately after
they're loaded), and it makes it less likely that we forget to report events. It
also lets us stop retaining Posts in Packages and gets us close to being able to
stop retaining Events as well.

Also cleans up a few test bugs this exposed ("WORKSPACE" as first line of test
workspace files).

PiperOrigin-RevId: 321144813