Add output_source_jar parameter to pack_sources Starlark function and allow empty jar files.
Before, by only having source_jar parameter, user in Starlark must actually have another Artifact (which is only used to determine the name).
With additional parameter output_source_jar it is possible to pack a jar and giving it a custom name.

Another problem was there was no way to create an empty .jar file, which is also a completely valid edge case.

RELNOTES: None.
PiperOrigin-RevId: 321344112