Remove unnecessary MinimalOutputStore class.
This is currently used as the store for an action filesystem, but it does not actually store any less data because injectOutputData is used to add data to the store, bypassing the overridden putArtifactData method. Instead, just use OutputStore for all use cases, and delete injectOutputData.

The original thought behind MinimalOutputStore was that we could reduce temporary caching when the action filesystem is in-memory anyway, but now OutputStore is down to just two maps required to construct the final ActionExecutionValue and there are no longer any temporary caches.

RELNOTES: None.
PiperOrigin-RevId: 321394749