Starlark: special case in `type` for list, tuple, dict, ...
`StarlarkInterfaceUtils.getStarlarkBuiltin` is quite expensive, and
at the same type `type` is usually invoked with simple types like
`list` or `dict`.

Results are:

```
A: N=7, r=5.492+-0.198
B: N=7, r=4.846+-0.354
B/A: 0.882
```

for for test:

```
def test():
    for i in range(10):
        print(i)
        for j in range(200):
            for k in range(1000):
                type([])
                type(())
                type({})
                type([])
                type(())
                type({})
                type([])
                type(())
                type({})
                type([])
                type(())
                type({})
                type([])
                type(())
                type({})

test()
```

For the invocations like `type(int)` the slowdown is about 1% but
less than the measurement noise.