Simplify creation of ActionMetadataHandler instances.
ActionMetadataHandler may need to be recreated after input discovery. This is messy because the constructor requires all of the fields again even though most are the same, and the fileset mapping is recomputed even though it is the same.

Instead of a single constructor, offer a static factory method for creating the initial handler, and a transformAfterInputDiscovery method with a single parameter for creating a new instance after input discovery.

RELNOTES: None.
PiperOrigin-RevId: 321446613