Remove the warning for shared actions with tree artifact inputs.
We recently added a warning whenever we encounter a shared action which has a
tree artifact input. The reason for it is because the shared action detection
may be wrong in such case and permit to share an action which actually is
different. That would lead to non-determinism and difficult to debug failures.

It turns out that shared actions with tree artifact inputs are quite
widespread, making the warning spammy. An example of it is
`java_proto_library`, which ends up creating a `GoogleJavaProtoAspect` with
`hasServices=false` and `hasServices=true`, both of which will create the same
`JavaCompileAction`.

Remove the user-facing warning and leave only blaze info log messages for
shared actions with tree artifact inputs.

PiperOrigin-RevId: 321474787