Add InstrumentedFilesInfo to Starlark globals
The documentation already listed that as one of the provider globals in the Starlark API. Adding that allows the provider to be read for Starlark, which is mostly useful for writing tests using analysistest that assert about the behavior of that provider.

(Note that this doesn't expose InstrumentedFilesInfo as a constructor. InstrumentedFilesInfo is constructed via coverage_common.instrumented_files_info, which configures that using provided lists of coverage-relevant source and dependency attributes. This just allows the provider to be looked up from a dependency, for the sake of testing. Basically, rules can return coverage_common.instrumented_files_info[...] and read dep[InstrumentedFilesInfo].)

RELNOTES: Add InstrumentedFilesInfo provider to Starlark globals.
PiperOrigin-RevId: 321578460