Move package loader results into custom business objects
This lets us add extra non-per-package info. The intention is to eventually
move events from being per-package to per-request. The existing way of getting
events per-package is potentially lossy (there can be other events besides those
attached to a package) and is based on an API we'd like to remove (package has
no reason to retain events associated with it besides for this code).

RELNOTES: None
PiperOrigin-RevId: 321640827