Refactor include validation logic into its own method
I a fixing a bug in ShowIncludesFilter [1]. For this I need to be able to invoke the include validation logic on a single artifact. Additionally, I'd consider this change a win for readability.

This change is a no-op. However, it might have implications on memory usage when loose header checking is used. This change always constructs the loose header checking set. I think this is fine as it's my understanding that loose header checking is barely used and going away.
[1] https://github.com/bazelbuild/bazel/issues/11765

RELNOTES: None
PiperOrigin-RevId: 321753605