Report all events from package loading in PackageLoader
There can be events besides those attached to packages. Moving forward, we're
looking to remove per-package event tracking, since it's completely unused
in bazel proper. While this is a loss of precision for PackageLoader, it is
more accurate (we get all of the events) and less likely to become outdated
(if we move events around).

PiperOrigin-RevId: 321776401