Support Label type as a parameter in repository_ctx.execute()
This change fixes an inconsistency between remotable and non-remotable
repository rules. If remotable, then repository_ctx.execute() accepts
label and string types as parameters. If non-remotable,
repository_ctx.execute() accepts string and path types as parameters.

An implementation written for a remotable rule should also work for
non-remotable rule. This change makes it so that a non-remotable
repository_rule also accepts a Label type as a parameter.

For context on remotable repository rules see
https://github.com/bazelbuild/bazel/commit/06a26c37126e3eefabd28c24a4e35059b5b136bc

Closes #11798.

PiperOrigin-RevId: 322544217