Split up actions java library target.
Notably, this separates Artifact and related types so that they don't have to be tied in with the rest of the actions library.

Pulled some nested classes to the top level to achieve this and did some other drive-by cleanups, but otherwise just import statements and BUILD files.

RELNOTES: None.
PiperOrigin-RevId: 322699250