When validating workers, also check that the worker hasn't died in the meanwhile.
Workers can die due to bugs in the tools or the wrappers or because something entirely different killed them. We cannot prevent that, but we can avoid trying to execute an action on a worker that's already dead.

This lets a build continue if a worker has died before it gets used for an action, but still fails the build if the worker dies during the action. Since the worker pool already has a check that newly created workers are valid, a worker that crashes will either get a failure from the worker pool, have an action fail, or make progress

RELNOTES: None
PiperOrigin-RevId: 322754088