bazel syntax: expose Resolver API
This change exposes a public API for accessing the results of name resolution.
All the concepts are defined by the spec.

Currently, the tree-based evaluator has privileged access to the Resolver
because it lives in the same Java package, but the package will soon be split
into two (syntax + eval).

When we switch to a compiled representation, there will be no need to access
resolver data structures at run-time. (Only the compiler need see them, and it
will live in the syntax package.) However, the resolver information may be useful
for a wide range of other tools that examine or refactor Starlark files.

PiperOrigin-RevId: 323358860