bazel packages: fix bug in CallStack prefix-sharing optimization
Prior to this change, the condition for detecting shared prefixes compared
corresponding frames but failed to check that all preceding frames also matched.
Consequently it could spuriously reuse nodes for a callstack with a common
frame in the middle but not at the start.

(The code was originally written as two loops, one to scan the common prefix
and one to create nodes for the different suffix. When I merged the two loops
because they were more similar than different, in effect I dropped the break
statement from the first loop.)

+ Test.

PiperOrigin-RevId: 323396274