bazel syntax: eliminate 'extends EvaluationTestCase' in this package
This change removes all but one use of inheritance of EvaluationTestCase,
instead favoring composition, which makes the flow of information explicit
and thus clearer.

A follow-up CL will fork EvaluationTestCase into to copies: a private
copy in lib/syntax, of which we can delete the bulk; and the public one,
which will be moved to lib/packages, on which it currently depends.

The sole remaining use of extends, in the FlagGuarding test, will be rewritten.

PiperOrigin-RevId: 323577917