Failed target/aspect completions include resolved artifact paths in BEP.
The implementation of feature request #9413 adds artifact URIs to BEP for action
groups that succeeded within an overall failed build (see commit ecd2a57662).

These artifact URIs were passed to the BES subsystem as local paths and bazel
relied on BES's artifact upload facility (which substitutes a remote URI after
upload) to make the artifacts available to other users.

With the introduction of "build without the bytes" functionality, in particular the
`--experimental_remote_download_outputs=minimal` setting, an artifact from a
successful action group may not be downloaded to the local filesystem. For targets
that succeed, the local path is converted to the remote URI *before* it is passed
to the BES subsystem which knows not to attempt an upload. Failed targets were
still passing an (erroneous) local path to BES, which would crash when it attempted
upload.

This change ensures failed targets whose artifacts have known remote URIs pass
those remote URIs to BES, which knows not to attempt an upload.

Fixes #11764.

PiperOrigin-RevId: 323581170