Let --remote_{cache,executor} support gRPC over UNIX sockets
All of the necessary code to construct gRPC clients that communicate
over UNIX sockets is there. It's just that we currently restrict its use
to the creation of proxies using --remote_proxy.

Let's shuffle some code around in GoogleAuthUtils to support the unix:
URL scheme directly. Move all of that code out of
newNettyChannelBuilder() into newUnixNettyChannelBuilder(), so that the
former can use it for both target and proxy channel creation.

UNIX sockets are useful, as they allow peers to authenticate using
light-weight features such as getsockopt(SO_PEERCRED).

Closes #11722.

PiperOrigin-RevId: 323771061