Handle event stdout/stderr as byte arrays, not strings.
This is to avoid trying to interpret the content of these streams in any
kind of encoding, because they may include binary data and we cannot assume
they don't.

We still have to interpret the output as strings when the flag
--experimental_ui_deduplicate is enabled, but this is disabled by default
and I'm tempted to say we should eliminate it in the name of simplicity.

RELNOTES: None.
PiperOrigin-RevId: 323794492