Do not expand implicit outputs of cc_library with pkg:*
We introduced a whitelist for removing the implicit outputs of cc_library, for targets not in the whitelist we add a provider which is detected during analysis if we are depending on the implicit output.

However, when doing blaze cquery build //somepackage:*, Bazel will still
list the implicit output and will try to analyze it directly.

To avoid this we stop listing the implicit output of cc_libraries while we go
through the remaining packages in the whitelist

RELNOTES:none
PiperOrigin-RevId: 323969068