Make PackageLoader implement AutoCloseable, and use the new #close method to shutdown internal threadpools (notably, the ForkJoinPool used for non-Skyframe globbing), thus fixing a resource leak.
This resource leak does NOT affect Bazel. If you don't know or care what PackageLoader is, you may stop reading.

The resource leak stems from a design flaw, which is my fault from the original PackageLoader CL many years ago. The previous lack of some sort of #close is evidence of my blatant oversight here.

The specific issue fixed by this CL is that we weren't shutting down the ForkJoinPool created by PackageFactory that is used for legacy globbing. The problem occurs when PackageLoader is asked to load a package with globs and high directory structure parallelism, thus enticing ForkJoinPool to create lots of internal threads. Then, since PackageLoader doesn't explicitly shutdown the ForkJoinPool, the threads (and various objects used for bookkeeping) stay around *after* the PackageLoader instance has been GC'd. This problem is largest for an app that uses short-lived PackageLoader instances---it's a memory leak. On the other hand, an app that e.g. uses a single long-lived PackageLoader instance would have no memory leak at all; this is the exact logical situation Bazel is in, since it a single PackageFactory instance whose lifetime is that of the Bazel server.

After this CL is submitted and sticks for a few business days, I will file an informational Github issue.

PiperOrigin-RevId: 324299253