Remove --experimental_ui_deduplicate.
This feature has caused confusion in the past (e.g. #7184) and is currently
disabled by default because of it.  Furthermore, this is conceptually broken
as it tries to parse stdout/stderr as line-oriented streams, but they
necessarily aren't, so any encoding conversions could result in errors.

In the name of simplicity, remove this experimental feature as I do not see
us wanting to make this the default again.

Fixes #7184.

RELNOTES[INC]: Removed the --experimental_ui_deduplicate flag.

PiperOrigin-RevId: 324605091