Add useDefaultShellEnvironment() in two places within the Java rules
Two files in the Java tools were not using useDefaultShellEnvironment()
when using the SpawnAction.Builder class. This causes the PATH to be dropped
when using these rules, even if the user passed --action_env=PATH. This is
problematic for me because I depend on the PATH to help Bazel find the MSVCRT
dlls.