Add blaze flag and support for splitting OPTIMIZATION bytecode optimizer action.
This adds a blaze flag ("split_bytecode_optimization_pass") to JavaOptions.
There is no functional difference this CL makes as long as the flag isn't
specified.
When the flag is specified, the OPTIMIZATION stage in the bytecode optimizer
will be split into 2 stages (OPTIMIZATION_INITIAL, OPTIMIZATION_FINAL),
each taking slightly over half the time of the original OPTIMIZATION stage.

PiperOrigin-RevId: 324702687