bazel syntax: include the Starlark stack in unhandled exceptions
This will help pinpoint the expression that causes a crash in the interpreter.

This applies to RuntimeException and StackOverflowError.

Bazel sometimes expects a particular unchecked exception to transit the
interpreter unchanged (e.g. MissingDepException). This is no longer true;
explicit getCause calls are now required.

Also, change tests not to assume a particular class of unchecked exception,
and add tests of NPE and StackOverflow.

PiperOrigin-RevId: 324817105