Catch exception instead of Throwable
This fixes the problem in #11579 but it might be a good idea to also change all occurences of Throwable to Exception in the class as discussed in the issue.

I created a new PR with a commit from the right email address to comply with the CLA (previous PR: #11580).

Ping: @laurentlb

Closes #11812.

PiperOrigin-RevId: 324975159