Clarify --instrumentation_filter documentation
This mostly changes "[a rule's] source files" to "the source files it processes" to be clear that we're talking about files belonging to the rule whose actions process those files, not the rule which _directly depends_ on the source file.

That's how --instrumentation_filter works because it lines up the "should you do something for coverage with these files at build time" logic and the "should you include these files in the manifest of instrumented files" logic. Wrapping a set of source files in a filegroup rule or similar doesn't change coverage semantics.

Related documentation about coverage_common.source_attributes is similarly updated.

RELNOTES: None.
PiperOrigin-RevId: 325267991