Fix parsing of BRDA lines in LcovParser
Fixes the interpretation of the "-" in a BRDA line (meaning the branch
was never executed, as opposed to executed but never taken) and fixes
the internal SourceFileCoverage abstraction to account for the fact that
multiple BRDA statements will occur for the same line.

Based on the PR https://github.com/bazelbuild/bazel/pull/10708 but does
not include any changes for the Java code coverage generator.

Fixes #11051

Closes #11899.

PiperOrigin-RevId: 325419922